Position and color vectors are integers after rotating

+++ Cubie.py
@@ -118,12 +118,12 @@ class Cubie(object):
 		else:
 			raise Exception(f"{axis} is not a valid axis! Must be x, y, or z!")
 
-		self.pos = np.round(R @ self.pos)
-		self.x = self.pos[0][0]
-		self.y = self.pos[1][0]
-		self.z = self.pos[2][0]
+		self.pos = np.round(R @ self.pos).astype(int)
+		self.x = int(self.pos[0][0])
+		self.y = int(self.pos[1][0])
+		self.z = int(self.pos[2][0])
 		for color, v  in self.color_to_dir.items():
-			self.color_to_dir[color] = np.round(R @ v)
+			self.color_to_dir[color] = np.round(R @ v).astype(int)
 
 	def __lt__(self, other):
 		'''
